Hotfix - webpack & bundle optimizations
