+++ .gitignore
@@ -28,4 +28,3 @@ distback/
 distCurrent/
 env/env.json
 env/.env
-
